25Jan
Thur

+++ canvas.h
@@ -24,7 +24,7 @@ class Canvas : public QOpenGLWidget, protected QOpenGLFunctions
 public:
     Canvas(QWidget *parent = 0);
 //  void updateGL();
-    void renderRectangle(int x, int y, int width, double height, bool boundary, QVector4D color);
+    void renderRectangle(float x, float y, float width, double height, bool boundary, QVector4D color);
        //QList<dataTuple>* listofStates = nullptr;
          QList<dataRegion>* regions = nullptr;
          QList<dataState>* states = nullptr;
